Move tests related to failure locations to own suite
